Fix navigation bug and replace debounce package
The bug when navigating search results using up/down arrow keys has been
fixed by moving the search results state into the parent component.
Something about the `AutoSuggest` component rerendering when the
highlighted item is changed that breaks the search results if it is
placed in the same component.

I also replaced the debounce package since the one from `usehooks-ts`
isn't working as intended, i.e., the debounced search would fire with
every keypress instead of only once when I'm finished typing my query.

Finally, I also added a message when there are no matching games to make
it clear for the user.
